# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Motewallet Withdrawal API
  description: 钱包提现系统 API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 开发环境
...
paths:
...
  /api/v1/admin/auth/login:
    post:
      tags: [Admin Auth]
      summary: 管理员登录
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: 登录成功
          headers:
            Set-Cookie:
              schema:
                type: string
                example: admin_token=eyJhbGciOi...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AdminInfo"
        "401":
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: 商户 JWT Token（httpOnly Cookie）
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: admin_token
      description: 管理员 JWT Token（httpOnly Cookie）
  schemas:
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 10001
        message:
          type: string
          example: invalid credentials
...
    AdminLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: superadmin
        password:
          type: string
          example: "Admin@2024"
...
    AdminInfo:
      type: object
      properties:
        id:
          type: integer
          format: uint64
          example: 1
        username:
          type: string
          example: superadmin
        email:
          type: string
          example: admin@motewallet.com
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN, OPERATOR, AUDITOR, VIEWER]
          example: SUPER_ADMIN
...